requirements.txt
easydict==1.9
Flask==2.1.0
matplotlib==3.3.4
numpy==1.19.5
opencv_python==4.5.3.56
Pillow==9.2.0
python-dotenv==0.20.0
requests==2.25.1
torch==1.12.0+cu116
torchvision==0.13.1
tqdm==4.62.3
